Ensure the gem works on Rails 4.2
On Rails 4.2, the ActiveSupport notification payload does not include
any `:type_casted_binds`. This was introduced in Rails 5.0. Thus when the
matcher does not match, it fails to display its message with the following
error:

    NoMethodError:
    undefined method `any?' for nil:NilClass

It happens while evaluating `query[:binds].any?` because `query[:binds]`
is `nil`.

This commit casts the values in `payload[:binds]` to have the same
behavior as with Rails 5.x

Also introduce appraisal to test from Rails 4.2 to 5.1

Also restricted Rails version to be '> 4.2' because specs fail with
Rails 4.1.